Create use-data-hook generator and use-project-data-hook
